+++ packages/odata2ts/src/generator/ImportContainer.ts
@@ -1,7 +1,7 @@
 import path from "path";
 
 import { ODataVersions } from "@odata2ts/odata-core";
-import { ImportDeclarationStructure } from "ts-morph";
+import { ImportDeclarationStructure, OptionalKind } from "ts-morph";
 
 import { DataModel } from "../data-model/DataModel";
 import { ComplexType, OperationType } from "../data-model/DataTypeModel";
@@ -16,7 +16,12 @@ import {
 } from "./import/ImportObjects";
 import { ImportedNameValidator } from "./ImportedNameValidator";
 
-type ImportContainerType = Record<keyof typeof LIB_MODULES, Map<string, string>>;
+type ImportContainerType = Record<
+  keyof typeof LIB_MODULES,
+  { regular: Map<string, string>; typeOnly: Map<string, string> }
+>;
+
+function generateImportId() {}
 
 /**
  * Handles all the import statements for a given file.
@@ -29,15 +34,21 @@ type ImportContainerType = Record<keyof typeof LIB_MODULES, Map<string, string>>
 export class ImportContainer {
   private readonly importedNameValidator: ImportedNameValidator;
   // mapping of a custom defined type to a primitive type
-  private customTypes = new Map<string, Map<string, string>>();
+  private customTypes = {
+    regular: new Map<string, Map<string, string>>(),
+    typeOnly: new Map<string, Map<string, string>>(),
+  };
   // imports to generated artefacts
-  private internalImports = new Map<string, Map<string, string>>();
+  private internalImports = {
+    regular: new Map<string, Map<string, string>>(),
+    typeOnly: new Map<string, Map<string, string>>(),
+  };
 
   private libs: ImportContainerType = {
-    core: new Map(),
-    qObject: new Map(),
-    clientApi: new Map(),
-    service: new Map(),
+    core: { regular: new Map(), typeOnly: new Map() },
+    qObject: { regular: new Map(), typeOnly: new Map() },
+    clientApi: { regular: new Map(), typeOnly: new Map() },
+    service: { regular: new Map(), typeOnly: new Map() },
   };
 
   constructor(
@@ -51,56 +62,62 @@ export class ImportContainer {
     this.importedNameValidator = new ImportedNameValidator(reservedNames);
   }
 
-  public addCoreLib(odataVersion: ODataVersions, ...coreLibs: Array<CoreImports>) {
-    return coreLibs.map((coreLib) => {
-      const name = CoreImports[coreLib] + (VERSIONED_CORE_IMPORTS.includes(coreLib) ? ODataVersions[odataVersion] : "");
-      const importName = this.importedNameValidator.validateName(LIB_MODULES.core, name);
-      this.libs.core.set(name, importName);
-      return importName;
-    });
+  public addCoreLib(odataVersion: ODataVersions, coreLib: CoreImports) {
+    const isVersioned = VERSIONED_CORE_IMPORTS.includes(coreLib);
+    const name = CoreImports[coreLib] + (isVersioned ? ODataVersions[odataVersion] : "");
+    const importName = this.importedNameValidator.validateName(LIB_MODULES.core, name);
+
+    // TODO: currently only types are imported, however enums could potentially be imported too
+    this.libs.core.typeOnly.set(name, importName);
+
+    return importName;
   }
 
-  public addFromQObject(...names: Array<string>) {
-    return names.map((n) => {
-      const importName = this.importedNameValidator.validateName(LIB_MODULES.qObject, n);
-      this.libs.qObject.set(n, importName);
-      return importName;
-    });
+  private addFromQObject(name: string, typeOnlyImport = false) {
+    const importName = this.importedNameValidator.validateName(LIB_MODULES.qObject, name);
+
+    const imports = typeOnlyImport ? this.libs.qObject.typeOnly : this.libs.qObject.regular;
+    imports.set(name, importName);
+
+    return importName;
   }
 
-  public addQObject(...qObjects: Array<QueryObjectImports>) {
-    return this.addFromQObject(...qObjects.map((qObject) => QueryObjectImports[qObject]));
+  public addQObject(qObject: QueryObjectImports | string) {
+    return this.addFromQObject(qObject);
   }
 
-  public addClientApi(...clientApis: Array<ClientApiImports>) {
-    return clientApis.map((clientApi) => {
-      const name = ClientApiImports[clientApi];
-      const importName = this.importedNameValidator.validateName(LIB_MODULES.clientApi, name);
-      this.libs.clientApi.set(name, importName);
-      return importName;
-    });
+  public addQObjectType(qObject: string) {
+    return this.addFromQObject(qObject, true);
   }
 
-  public addServiceObject(odataVersion: ODataVersions, ...serviceObjects: Array<ServiceImports>) {
-    return serviceObjects
-      .map((so) => {
-        return ServiceImports[so] + (VERSIONED_SERVICE_IMPORTS.includes(so) ? ODataVersions[odataVersion] : "");
-      })
-      .map((so) => {
-        const importName = this.importedNameValidator.validateName(LIB_MODULES.service, so);
-        this.libs.service.set(so, importName);
-        return importName;
-      });
+  public addClientApi(clientApi: ClientApiImports) {
+    const name = ClientApiImports[clientApi];
+    const importName = this.importedNameValidator.validateName(LIB_MODULES.clientApi, name);
+
+    // complete client api consists only of types
+    this.libs.clientApi.typeOnly.set(name, importName);
+    return importName;
   }
 
-  public addCustomType(moduleName: string, typeName: string) {
-    let importList = this.customTypes.get(moduleName);
+  public addServiceObject(odataVersion: ODataVersions, serviceObject: ServiceImports) {
+    const isVersioned = VERSIONED_SERVICE_IMPORTS.includes(serviceObject);
+    const name = ServiceImports[serviceObject] + (isVersioned ? ODataVersions[odataVersion] : "");
+    const importName = this.importedNameValidator.validateName(LIB_MODULES.service, name);
+
+    // only regular imports for the service package
+    this.libs.service.regular.set(name, importName);
+    return importName;
+  }
+
+  public addCustomType(moduleName: string, typeName: string, isTypeOnly: boolean = false) {
+    const importName = this.importedNameValidator.validateName(moduleName, typeName);
+    const imports = isTypeOnly ? this.customTypes.typeOnly : this.customTypes.regular;
+    let importList = imports.get(moduleName);
     if (!importList) {
       importList = new Map();
-      this.customTypes.set(moduleName, importList);
+      imports.set(moduleName, importList);
     }
 
-    const importName = this.importedNameValidator.validateName(moduleName, typeName);
     importList.set(typeName, importName);
     return importName;
   }
@@ -113,7 +130,7 @@ export class ImportContainer {
     return this.pathAndFile(this.path, this.fileName) !== this.pathAndFile(filePath, fileName);
   }
 
-  private addGeneratedImport(folderPath: string, fileName: string, name: string): string {
+  private addGeneratedImport(folderPath: string, fileName: string, name: string, isTypeOnly = false): string {
     // imports are only relevant for different files
     if (!this.isDifferentFile(folderPath, fileName)) {
       return name;
@@ -121,17 +138,18 @@ export class ImportContainer {
 
     const moduleName = this.pathAndFile(folderPath, fileName);
     const importName = this.importedNameValidator.validateName(moduleName, name);
+    const imports = isTypeOnly ? this.internalImports.typeOnly : this.internalImports.regular;
 
-    const imports = this.internalImports.get(moduleName) || new Map<string, string>();
-    imports.set(name, importName);
-    this.internalImports.set(moduleName, imports);
+    const importList = imports.get(moduleName) || new Map<string, string>();
+    importList.set(name, importName);
+    imports.set(moduleName, importList);
 
     return importName;
   }
 
   public addGeneratedModel(fqName: string, name: string): string {
     if (this.bundledFileGeneration) {
-      return this.addGeneratedImport("", this.mainFileNames.model, name);
+      return this.addGeneratedImport("", this.mainFileNames.model, name, true);
     } else {
       const model = this.dataModel.getModel(fqName) as ComplexType | undefined;
       if (!model && fqName !== "") {
@@ -140,11 +158,11 @@ export class ImportContainer {
 
       const folderPath = model ? model.folderPath : "";
       const modelName = model ? model.modelName : this.mainFileNames.model;
-      return this.addGeneratedImport(folderPath, modelName, name);
+      return this.addGeneratedImport(folderPath, modelName, name, true);
     }
   }
 
-  public addGeneratedQObject(fqName: string, name: string) {
+  public addGeneratedQObject(fqName: string, name: string, isTypeOnly = false) {
     if (this.bundledFileGeneration) {
       return this.addGeneratedImport("", this.mainFileNames.qObject, name);
     } else {
@@ -155,7 +173,7 @@ export class ImportContainer {
 
       const folderPath = model ? model.folderPath : "";
       const qName = model ? model.qName : this.mainFileNames!.qObject;
-      return this.addGeneratedImport(folderPath, qName, name);
+      return this.addGeneratedImport(folderPath, qName, name, isTypeOnly);
     }
   }
 
@@ -168,32 +186,54 @@ export class ImportContainer {
     }
   }
 
-  public getImportDeclarations(fromSubPath: boolean = false): Array<ImportDeclarationStructure> {
+  private createImportDecl(
+    module: string,
+    toImport: Map<string, string>,
+    isTypeOnly = false
+  ): OptionalKind<ImportDeclarationStructure> {
+    return {
+      namedImports: this.getNamedImports(toImport),
+      moduleSpecifier: module,
+      isTypeOnly,
+    };
+  }
+
+  public getImportDeclarations(): Array<OptionalKind<ImportDeclarationStructure>> {
     return [
-      ...Object.entries(this.libs)
-        .filter(([_, values]) => !!values.size)
+      ...Object.entries(this.libs).reduce<Array<OptionalKind<ImportDeclarationStructure>>>(
+        (result, [moduleName, toImport]) => {
+          const module = LIB_MODULES[moduleName as keyof typeof LIB_MODULES];
+          if (toImport.typeOnly.size) {
+            result.push(this.createImportDecl(module, toImport.typeOnly, true));
+          }
+          if (toImport.regular.size) {
+            result.push(this.createImportDecl(module, toImport.regular));
+          }
+          return result;
+        },
+        []
+      ),
+      ...[...this.customTypes.typeOnly]
+        .filter(([moduleName, toImport]) => toImport.size > 0)
         .map(([moduleName, toImport]) => {
-          return {
-            namedImports: this.getNamedImports(toImport),
-            moduleSpecifier: LIB_MODULES[moduleName as keyof typeof LIB_MODULES],
-          } as ImportDeclarationStructure;
+          return this.createImportDecl(moduleName, toImport, true);
         }),
-      ...[...this.customTypes.keys()]
-        .filter((moduleName) => !!this.customTypes.get(moduleName)?.size)
-        .map((moduleName) => {
-          const toImport = this.customTypes.get(moduleName)!;
-          return {
-            namedImports: this.getNamedImports(toImport),
-            moduleSpecifier: moduleName,
-          } as ImportDeclarationStructure;
+      ...[...this.customTypes.regular]
+        .filter(([moduleName, toImport]) => toImport.size > 0)
+        .map(([moduleName, toImport]) => {
+          return this.createImportDecl(moduleName, toImport);
         }),
-      ...[...this.internalImports.entries()]
+      ...[...this.internalImports.typeOnly]
         .filter(([_, toImport]) => toImport.size > 0)
         .map(([key, toImport]) => {
-          return {
-            namedImports: this.getNamedImports(toImport),
-            moduleSpecifier: this.getModuleSpecifier(key),
-          } as ImportDeclarationStructure;
+          const module = this.getModuleSpecifier(key);
+          return this.createImportDecl(module, toImport, true);
+        }),
+      ...[...this.internalImports.regular]
+        .filter(([_, toImport]) => toImport.size > 0)
+        .map(([key, toImport]) => {
+          const module = this.getModuleSpecifier(key);
+          return this.createImportDecl(module, toImport);
         }),
     ];
   }