+++ packages/odata2ts/test/fixture/generator/model/entity-converter-with-model.ts
@@ -1,4 +1,4 @@
-import { PrefixModel } from "@odata2ts/test-converters";
+import type { PrefixModel } from "@odata2ts/test-converters";
 
 export interface Book {
   id: boolean;