+++ packages/odata2ts/test/fixture/generator/qobject/action-bound.ts
@@ -1,7 +1,7 @@
 import { QAction, QBooleanPath, QStringParam, QueryObject } from "@odata2ts/odata-query-objects";
 
 // @ts-ignore
-import { Book_BoundActionParams } from "./TesterModel";
+import type { Book_BoundActionParams } from "./TesterModel";
 
 export class QBook extends QueryObject {
   public readonly id = new QBooleanPath(this.withPrefix("id"));