+++ packages/odata2ts/test/fixture/generator/qobject/function-min.ts
@@ -1,7 +1,7 @@
 import { QFunction, QStringParam } from "@odata2ts/odata-query-objects";
 
 // @ts-ignore
-import { MinFunctionParams } from "./TesterModel";
+import type { MinFunctionParams } from "./TesterModel";
 
 export class QMinFunction extends QFunction<MinFunctionParams> {
   private readonly params = [new QStringParam("test"), new QStringParam("optTest")];