+++ packages/odata2ts/test/fixture/generator/service/v4/big-number-return-types.ts
@@ -1,5 +1,5 @@
-import { HttpResponseModel, ODataHttpClient, ODataHttpClientConfig } from "@odata2ts/http-client-api";
-import { ODataCollectionResponseV4, ODataValueResponseV4 } from "@odata2ts/odata-core";
+import type { HttpResponseModel, ODataHttpClient, ODataHttpClientConfig } from "@odata2ts/http-client-api";
+import type { ODataCollectionResponseV4, ODataValueResponseV4 } from "@odata2ts/odata-core";
 import { EntitySetServiceV4, EntityTypeServiceV4, ODataService } from "@odata2ts/odata-service";
 
 // @ts-ignore
@@ -12,7 +12,7 @@ import {
   qTestEntity,
 } from "./QTester";
 // @ts-ignore
-import { EditableTestEntity, TestEntity, TestEntityId } from "./TesterModel";
+import type { EditableTestEntity, TestEntity, TestEntityId } from "./TesterModel";
 
 export class TesterService<in out ClientType extends ODataHttpClient> extends ODataService<ClientType> {
   private _qPingBigNumber?: QPingBigNumber;