+++ packages/odata2ts/test/fixture/generator/service/v4/big-numbers.ts
@@ -1,11 +1,12 @@
-import { ODataHttpClient } from "@odata2ts/http-client-api";
-import { QBigNumberCollection, StringCollection, qBigNumberCollection } from "@odata2ts/odata-query-objects";
+import type { ODataHttpClient } from "@odata2ts/http-client-api";
+import type { QBigNumberCollection, StringCollection } from "@odata2ts/odata-query-objects";
+import { qBigNumberCollection } from "@odata2ts/odata-query-objects";
 import { CollectionServiceV4, EntitySetServiceV4, EntityTypeServiceV4, ODataService } from "@odata2ts/odata-service";
 
 // @ts-ignore
 import { QTestEntity, QTestEntityId, qTestEntity } from "./QTester";
 // @ts-ignore
-import { EditableTestEntity, TestEntity, TestEntityId } from "./TesterModel";
+import type { EditableTestEntity, TestEntity, TestEntityId } from "./TesterModel";
 
 export class TesterService<in out ClientType extends ODataHttpClient> extends ODataService<ClientType> {
   constructor(client: ClientType, basePath: string) {