+++ packages/odata2ts/test/fixture/generator/service/v4/function-bound-unbound.ts
@@ -1,11 +1,11 @@
-import { HttpResponseModel, ODataHttpClient, ODataHttpClientConfig } from "@odata2ts/http-client-api";
-import { ODataCollectionResponseV4, ODataModelResponseV4 } from "@odata2ts/odata-core";
+import type { HttpResponseModel, ODataHttpClient, ODataHttpClientConfig } from "@odata2ts/http-client-api";
+import type { ODataCollectionResponseV4, ODataModelResponseV4 } from "@odata2ts/odata-core";
 import { EntitySetServiceV4, EntityTypeServiceV4, ODataService } from "@odata2ts/odata-service";
 
 // @ts-ignore
 import { QFirstBook, QGetBestsellers, QTestEntity, QTestEntityId, qTestEntity } from "./QTester";
 // @ts-ignore
-import { EditableTestEntity, FirstBookParams, TestEntity, TestEntityId } from "./TesterModel";
+import type { EditableTestEntity, FirstBookParams, TestEntity, TestEntityId } from "./TesterModel";
 
 export class TesterService<in out ClientType extends ODataHttpClient> extends ODataService<ClientType> {
   private _qGetBestsellers?: QGetBestsellers;