+++ examples/odata-v2/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.12](https://github.com/odata2ts/odata2ts/compare/@odata2ts/example-odata-v2@0.5.11...@odata2ts/example-odata-v2@0.5.12) (2024-07-31)
+
+**Note:** Version bump only for package @odata2ts/example-odata-v2
+
+## [0.5.11](https://github.com/odata2ts/odata2ts/compare/@odata2ts/example-odata-v2@0.5.10...@odata2ts/example-odata-v2@0.5.11) (2024-07-31)
+
+**Note:** Version bump only for package @odata2ts/example-odata-v2
+
 ## [0.5.10](https://github.com/odata2ts/odata2ts/compare/@odata2ts/example-odata-v2@0.5.9...@odata2ts/example-odata-v2@0.5.10) (2024-05-16)
 
 ### Bug Fixes